pyproject.toml
[project]
name = "text2sql"
version = "0.1.0"
description = "Document-grounded text-to-SQL evaluation and training"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "torch>=2.5",
  "transformers>=4.48",
  "datasets>=3.0",
  "accelerate>=1.0",
  "peft>=0.14",
  "trl>=0.15",
  "vllm>=0.7",
  "pydantic>=2.8",
  "typer>=0.12",
  "rich>=13.7",
  "orjson>=3.10",
  "polars>=1.8",
  "pyarrow>=18.0",
  "huggingface-hub>=0.28",
  "llm-os-eval-core",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3",
  "pytest-xdist>=3.6",
  "ipykernel>=6.29",
  "matplotlib>=3.9",
  "pandas>=2.2",
  "ruff>=0.8",
]

[tool.uv.sources]
llm-os-eval-core = { path = "../llm-os-eval-core" }

[tool.uv]
